# [perf] Faster to_string for all resource identifiers