--- a/crates/types/Cargo.toml
+++ b/crates/types/Cargo.toml
@@ -46,6 +46,7 @@ enum_dispatch = "0.3.13"
 enumset = { workspace = true, features = ["serde"] }
 figment = { version = "0.10.8", features = ["env", "toml"] }
 flexbuffers = { workspace = true }
+generic-array = { version = "1.2.0" }
 hostname = { workspace = true }
 http = { workspace = true }
 humantime = { workspace = true }