--- a/crates/types/benches/id_encoding.rs
+++ b/crates/types/benches/id_encoding.rs
@@ -12,6 +12,7 @@ use std::fmt::Write;
 
 use criterion::{Criterion, criterion_group, criterion_main};
 
+use restate_types::identifiers::AwakeableIdentifier;
 use restate_types::{IdEncoder, identifiers::InvocationId};
 
 pub fn id_encoding(c: &mut Criterion) {
@@ -23,6 +24,25 @@ pub fn id_encoding(c: &mut Criterion) {
                 buf.clear();
                 write!(&mut buf, "{id}")
             },
+            criterion::BatchSize::SmallInput,
+        );
+    })
+    .bench_function("invocation-id-to_string", |b| {
+        b.iter_batched(
+            InvocationId::mock_random,
+            |id| id.to_string(),
+            criterion::BatchSize::SmallInput,
+        );
+    })
+    .bench_function("awakeable-id-to_string", |b| {
+        b.iter_batched(
+            || {
+                let invocation_id = InvocationId::mock_random();
+                let entry_index = rand::random();
+
+                AwakeableIdentifier::new(invocation_id, entry_index)
+            },
+            |id| id.to_string(),
             criterion::BatchSize::SmallInput,
         );
     });