--- a/crates/types/src/lib.rs
+++ b/crates/types/src/lib.rs
@@ -55,7 +55,7 @@ pub mod storage;
 pub mod time;
 pub mod timer;
 
-pub use id_util::{IdDecoder, IdEncoder, IdResourceType, IdStrCursor};
+pub use id_util::IdResourceType;
 pub use node_id::*;
 pub use restate_version::*;
 pub use version::*;