Allocation free to_string and Display impls for all resource identifiers. While the previous version was great for pre-allocated buffers to Fmt into, it could perform better in `to_string()` scenario by being fully allocation free.

```console
~/w/r/restate ❯❯❯ cargo bench -p restate-types  --bench id_encoding
    Finished `bench` profile [optimized + debuginfo] target(s) in 55.61s
     Running benches/id_encoding.rs (target/release/deps/id_encoding-e2f9e9fd3134760e)
invocation-id-display   time:   [38.700 ns 38.745 ns 38.798 ns]
                        change: [-11.835% -11.510% -11.170%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  3 (3.00%) low severe
  1 (1.00%) low mild
  7 (7.00%) high mild
  2 (2.00%) high severe

invocation-id-to_string time:   [44.330 ns 44.712 ns 45.320 ns]
                        change: [-55.251% -54.827% -54.181%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild
  3 (3.00%) high severe

awakeable-id-to_string  time:   [25.942 ns 26.095 ns 26.263 ns]
                        change: [-74.492% -74.147% -73.829%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  5 (5.00%) high mild
  4 (4.00%) high severe

```