@peterbourgon 2025-08-28
> Running benches/id_encoding.rs (target/release/deps/id_encoding-5d457b0262b9c4cb) ...

Just to satisfy my curiosity, could you show the command(s) you used to get this output, in a way that I could reproduce?