@AhmedSoliman 2025-08-28
> ```
> cargo bench -p restate-types  --bench id_encoding
> ```

@peterbourgon I've updated the description to include the command. The flow would be to first navigate to the first commit in this PR, it only introduces the benchmark. Then run:
```cargo bench -p restate-types  --bench id_encoding```

After it prints the results, navigate to the next commit and run the same command again.