@AhmedSoliman 2025-08-28
@jackkleeman I've included a couple of other optimizations, this should be ready now.